Tilemap WIP + Deleted "Stagedbuilds"
Tilemap done besides Cave Background.

Removed "WindowsNoEditor" from StagedBuilds in "UnrealProjects">"Project_01">"Saved">"StagedBuilds" as it prevented uploading to the online repository. Unsure if this will cause future issues, but if it does I have the file saved in "Pictures" just in case I need to put it back. Hopefully its not loadbearing in any way but we will see! :)