Updated TileMap, Created Collision barrier, Created Kill Barrier, Created Enum, and Added Healthbar.
Updated Tilemap as it was not setup properly for the fixed jump.

Created Enum to route PlayerCharacter movement through.

Created Healthbar via widget, added functionality that accurately reflects PlayerCharacter's current health.

Added an invisible collision barrier around the beginning of the map to prevent PlayerCharacter from escaping out of bounds.

Created & added a "Kill Barrier" that resets the player to the start of the level upon falling out of the map.

Began working on adding a "Damage" & "Death" state for PlayerCharacter.